Last changes from security

+++ src/main/java/com/api/estramipyme/respositories/UserRepository.java
@@ -1,10 +1,11 @@
 package com.api.estramipyme.respositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.security.core.userdetails.User;
+import com.api.estramipyme.user.User;
+
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String email);
 }
